PartTimeJobApi/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cloudinary==1.44.1
Django==6.0.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.17.1
mysqlclient==2.2.8
packaging==26.0
pip-review==1.3.0
PyMySQL==1.1.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3